add more functionality to common file and clean up stock lots report.
